fix(layering): move benchmark-class helpers to ports so infra can import safely
TestLayeringRules enforces that internal/infra must not import domain
logic (except domain/*/ports and domain/vod). runtime_hardening.go was
importing internal/domain/playbackprofile only for BenchmarkClassForProfile.

Move BenchmarkClassForCodec + BenchmarkClassForProfile into the existing
playbackprofile/ports package (which already owns HostBenchmarkSnapshot
and the profile-ID constants).  The domain package re-exports them as
thin wrappers so every existing control-layer caller stays unchanged.

Fixes #563 — CI Nightly run 27114338154 layering failure:
  internal/infra/media/ffmpeg/runtime_hardening.go imports
  github.com/ManuGH/xg2g/internal/domain/playbackprofile
